Fix ongoing breakages in pip builds
1. CL/161014271 added timeseries to tensorflow, but did not add the files to the TensorFlow pip package, resulting in Jenkins test errors: e.g., http://ci.tensorflow.org/job/tensorflow-cl-cpu-python3-pip/9269/console

2. Fix numpy issues related to using float as size args, which is illegal in some recent versions of numpy.

3. The tensorflow/contrib/boosted_trees py_tests are not working in the pip (specifically bazel_pip) builds. Disable them in pip builds for now after filing b/63386583.

PiperOrigin-RevId: 161097246

+++ tensorflow/contrib/timeseries/python/timeseries/ar_model_test.py
@@ -66,7 +66,7 @@ class ARModelTest(test.TestCase):
              noise_stddev / 3. * np.random.randn(num_samples, 1))
     # Add some anomalies to data1
     if anomaly_prob > 0.:
-      num_anomalies = anomaly_prob * num_samples
+      num_anomalies = int(anomaly_prob * num_samples)
       anomaly_values = (anomaly_stddev_scale * noise_stddev / 4 *
                         np.random.randn(num_anomalies))
       indices = np.random.randint(0, num_samples, num_anomalies)
